+++ refena/lib/src/provider/types/future_provider.dart
@@ -3,16 +3,12 @@ import 'package:refena/src/async_value.dart';
 import 'package:refena/src/notifier/base_notifier.dart';
 import 'package:refena/src/provider/base_provider.dart';
 import 'package:refena/src/provider/override.dart';
-import 'package:refena/src/provider/types/async_notifier_provider.dart';
 import 'package:refena/src/provider/types/future_family_provider.dart';
 import 'package:refena/src/proxy_ref.dart';
 import 'package:refena/src/ref.dart';
 
-/// A [FutureProvider] is a custom implementation of a [AsyncNotifierProvider]
-/// that allows you to watch a [Future].
-///
-/// The advantage over using a [FutureBuilder] (by Flutter) is that the
-/// value is cached and only the first call to the [Future] is executed.
+/// A [FutureProvider] allows you to watch a [Future] in your widgets.
+/// The value is cached and only fetched once.
 ///
 /// Usage:
 /// final myProvider = FutureProvider((ref) async {