+++ refena/lib/src/provider/types/stream_family_provider.dart
@@ -3,29 +3,29 @@ import 'package:refena/src/async_value.dart';
 import 'package:refena/src/notifier/base_notifier.dart';
 import 'package:refena/src/provider/base_provider.dart';
 import 'package:refena/src/provider/override.dart';
-import 'package:refena/src/provider/types/future_provider.dart';
+import 'package:refena/src/provider/types/stream_provider.dart';
 import 'package:refena/src/provider/watchable.dart';
 import 'package:refena/src/ref.dart';
 
-typedef FutureFamilyBuilder<T, P> = Future<T> Function(
+typedef StreamFamilyBuilder<T, P> = Stream<T> Function(
   WatchableRef ref,
   P param,
 );
 
-/// A [FutureFamilyProvider] is a special version of [FutureProvider] that
-/// allows you to watch a collection of [Future]s.
-class FutureFamilyProvider<T, P> extends BaseProvider<
+/// A [StreamFamilyProvider] is a special version of [StreamProvider] that
+/// allows you to watch a collection of [Stream]s.
+class StreamFamilyProvider<T, P> extends BaseProvider<
     FamilyNotifier<AsyncValue<T>, P>, Map<P, AsyncValue<T>>> {
-  FutureFamilyProvider(
+  StreamFamilyProvider(
     this._builder, {
     super.onChanged,
     String Function(AsyncValue<T> state)? describeState,
     String? debugLabel,
     super.debugVisibleInGraph = true,
   })  : _describeState = describeState,
-        super(debugLabel: debugLabel ?? 'FutureFamilyProvider<$T, $P>');
+        super(debugLabel: debugLabel ?? 'StreamFamilyProvider<$T, $P>');
 
-  final FutureFamilyBuilder<T, P> _builder;
+  final StreamFamilyBuilder<T, P> _builder;
 
   final String Function(AsyncValue<T> state)? _describeState;
 
@@ -35,12 +35,12 @@ class FutureFamilyProvider<T, P> extends BaseProvider<
     return _buildFamilyNotifier(this, _builder, _describeState);
   }
 
-  /// Overrides the future builder.
+  /// Overrides the stream builder.
   ///
   /// {@category Initialization}
   ProviderOverride<FamilyNotifier<AsyncValue<T>, P>, Map<P, AsyncValue<T>>>
-      overrideWithFutureBuilder(
-    FutureFamilyBuilder<T, P> builder,
+      overrideWithStreamBuilder(
+    StreamFamilyBuilder<T, P> builder,
   ) {
     return ProviderOverride(
       provider: this,
@@ -57,12 +57,12 @@ class FutureFamilyProvider<T, P> extends BaseProvider<
 }
 
 FamilyNotifier<AsyncValue<T>, P> _buildFamilyNotifier<T, P>(
-  FutureFamilyProvider<T, P> provider,
-  FutureFamilyBuilder<T, P> builder,
+  StreamFamilyProvider<T, P> provider,
+  StreamFamilyBuilder<T, P> builder,
   String Function(AsyncValue<T> state)? describeState,
 ) {
   return FamilyNotifier<AsyncValue<T>, P>(
-    (param) => FutureProvider<T>(
+    (param) => StreamProvider<T>(
       (ref) => builder(ref, param),
       debugLabel: '${provider.debugLabel}($param)',
     ),