+++ refena/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.4.0
 
 - feat: make `FutureProvider` and `StreamProvider` rebuildable (`ref.watch` is allowed)
+- feat: add family version of `StreamProvider`
 - feat: add `Ref.rebuild` to manually rebuild a rebuildable provider
 
 ## 1.3.0